+++ src/L01E08_Imagem/structs/Pixel/Pixel.hpp
@@ -6,30 +6,30 @@
 
 class Pixel {
     private:
-        unsigned short int red;
-        unsigned short int green;
-        unsigned short int blue;
+        int red;
+        int green;
+        int blue;
     
     public:
         Pixel();
-        Pixel(const unsigned short int red, const unsigned short int green, const unsigned short int blue);
+        Pixel(const int red, const int green, const int blue);
 
         void print() const;
 
         void set_rgb(const std::string& cor);
-        void set_rgb(const unsigned short int red, const unsigned short int green, const unsigned short int blue);
-        std::vector<unsigned short int> get_rgb() const;
+        void set_rgb(const int red, const int green, const int blue);
+        std::vector<int> get_rgb() const;
 
-        void set_red(const unsigned short int red);
-        unsigned short int get_red() const;
+        void set_red(const int red);
+        int get_red() const;
 
-        void set_green(const unsigned short int green);
-        unsigned short int get_green() const;
+        void set_green(const int green);
+        int get_green() const;
 
-        void set_blue(const unsigned short int blue);
-        unsigned short int get_blue() const;
+        void set_blue(const int blue);
+        int get_blue() const;
     
-        unsigned short int get_tom_medio() const;
+        int get_tom_medio() const;
 };
 
 #endif
